+++ packages/react-ui/src/theme/themeUtils.js
@@ -1,3 +1,4 @@
+import ReactDOMServer from 'react-dom/server';
 import { createSpacing } from '@mui/system';
 import { SPACING } from './themeConstants';
 
@@ -10,4 +11,14 @@ export function getPixelToRem(px) {
   const rem = (1 / fontBase) * px + 'rem';
 
   return rem;
+}
+
+// Get the icon path from a given icon
+export function getIconPath(icon) {
+  const iconString = ReactDOMServer.renderToString(icon);
+  const parser = new DOMParser();
+  const svg = parser.parseFromString(iconString, 'image/svg+xml');
+  const iconPath = svg.querySelector('path').getAttribute('d');
+
+  return iconPath;
 }